requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1